refactor / move waiting for activation updates
