+++ validator/client/runner_test.go
@@ -206,55 +206,3 @@ func TestAllValidatorsAreExited_NextSlot(t *testing.T) {
 	run(ctx, v)
 	assert.LogsContain(t, hook, "All validators are exited")
 }
-
-func TestHandleAccountsChanged_Ok(t *testing.T) {
-	ctx := context.Background()
-	defer ctx.Done()
-
-	km := &mockKeymanager{accountsChangedFeed: &event.Feed{}}
-	v := &FakeValidator{Keymanager: km}
-	channel := make(chan struct{})
-	go handleAccountsChanged(ctx, v, channel)
-	time.Sleep(time.Second) // Allow time for subscribing to changes.
-	km.SimulateAccountChanges()
-	time.Sleep(time.Second) // Allow time for handling subscribed changes.
-
-	select {
-	case _, ok := <-channel:
-		if !ok {
-			t.Error("Account changed channel is closed")
-		}
-	default:
-		t.Error("Accounts changed channel is empty")
-	}
-}
-
-func TestHandleAccountsChanged_CtxCancelled(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-
-	km := &mockKeymanager{accountsChangedFeed: &event.Feed{}}
-	v := &FakeValidator{Keymanager: km}
-	channel := make(chan struct{}, 2)
-	go handleAccountsChanged(ctx, v, channel)
-	time.Sleep(time.Second) // Allow time for subscribing to changes.
-	km.SimulateAccountChanges()
-	time.Sleep(time.Second) // Allow time for handling subscribed changes.
-
-	cancel()
-	time.Sleep(time.Second) // Allow time for handling cancellation.
-	km.SimulateAccountChanges()
-	time.Sleep(time.Second) // Allow time for handling subscribed changes.
-
-	var values int
-	for loop := true; loop == true; {
-		select {
-		case _, ok := <-channel:
-			if ok {
-				values++
-			}
-		default:
-			loop = false
-		}
-	}
-	assert.Equal(t, 1, values, "Incorrect number of values were passed to the channel")
-}
