+++ validator/client/validator_test.go
@@ -93,6 +93,9 @@ func (m *mockKeymanager) Sign(ctx context.Context, req *validatorpb.SignRequest)
 }
 
 func (m *mockKeymanager) SubscribeAccountChanges(pubKeysChan chan [][48]byte) event.Subscription {
+	if m.accountsChangedFeed == nil {
+		m.accountsChangedFeed = &event.Feed{}
+	}
 	return m.accountsChangedFeed.Subscribe(pubKeysChan)
 }
 
@@ -362,7 +365,7 @@ func TestWaitMultipleActivation_LogsActivationEpochOK(t *testing.T) {
 		resp,
 		nil,
 	)
-	require.NoError(t, v.WaitForActivation(context.Background(), make(chan struct{})), "Could not wait for activation")
+	require.NoError(t, v.WaitForActivation(context.Background()), "Could not wait for activation")
 	require.LogsContain(t, hook, "Validator activated")
 }
 
@@ -400,7 +403,7 @@ func TestWaitActivation_NotAllValidatorsActivatedOK(t *testing.T) {
 		resp,
 		nil,
 	)
-	assert.NoError(t, v.WaitForActivation(context.Background(), make(chan struct{})), "Could not wait for activation")
+	assert.NoError(t, v.WaitForActivation(context.Background()), "Could not wait for activation")
 }
 
 func TestWaitSync_ContextCanceled(t *testing.T) {
